chore: Components 디렉토리 생성 및 파일 이동
