+++ README.md
@@ -19,16 +19,12 @@
 
 - [x] small, medium, large 등의 크기 옵션을 prop으로 전달받아 모달 크기 조절
 
-### 3. 반응형 구현
-
-- [ ] PC 환경에서의 사용을 고려하여 대응하기(반응형 디자인 확장)
-
-### 4. 모달 컴포넌트 웹 접근성(A11y) 개선
+### 3. 모달 컴포넌트 웹 접근성(A11y) 개선
 
 - [x] 키보드 사용자를 고려하여 모달 내부 포커스 제어 구현하기
-- [x] 모달이 열릴 때 모달 내부의 첫 번째 포커스 가능한 요소로 자동 포커스를 이동시키기
-- [x] TAB 키로 모달 내부의 포커스 가능한 요소를 순서대로 탐색할 수 있도록 구현하기
-- [ ] Shift+TAB 키로 역순 탐색이 가능해야 하며, 마지막 요소에서 첫 번째 요소로 순환 가능하도록 구현하기
+  - [x] 모달이 열릴 때 모달 내부의 첫 번째 포커스 가능한 요소로 자동 포커스를 이동시키기
+  - [x] TAB 키로 모달 내부의 포커스 가능한 요소를 순서대로 탐색할 수 있도록 구현하기
+  - [ ] Shift+TAB 키로 역순 탐색이 가능해야 하며, 마지막 요소에서 첫 번째 요소로 순환 가능하도록 구현하기
 
 ## [Step1] ⚙️ Module
