--- components/src/lib/Modal/index.ts
@@ -1,3 +1,0 @@
-export { default as AlertModal } from './AlertModal';
-export { default as ConfirmModal } from './ConfirmModal';
-export { default as PromptModal } from './PromptModal';
