+++ components/src/lib/components/Button/index.ts
@@ -1,0 +1,1 @@
+export { default as Button } from './Button.tsx';
