+++ components/src/lib/components/Modal/PromptModal.tsx
@@ -1,4 +1,4 @@
-import Modal from '../Modal';
+import Modal from '../../Modal.tsx';
 import styled from '@emotion/styled';
 import { Button } from "../Button";
 import { useState } from 'react';
