+++ components/src/lib/components/Modal/index.ts
@@ -1,0 +1,3 @@
+export { default as AlertModal } from './AlertModal.tsx';
+export { default as ConfirmModal } from './ConfirmModal.tsx';
+export { default as PromptModal } from './PromptModal.tsx';
